requirements.txt
word2number==1.1
pint==0.24.4
dateparser==1.2.1
inflect==7.5.0
faiss-cpu==1.11.0
spacy==3.8.7
nltk==3.9.1
sentence-transformers==4.1.0
